requirements.txt
openai
python-dotenv


